# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Customers",
        "description": "Documentation de l'API REST du webservice Customers du projet PayeTonKawa pour la MSPR TPRE814.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List all customers",
                "description": "Display a listing of the resource.",
                "operationId": "6f01ddd63e2a72fee24eb31157a1c799",
                "responses": {
                    "200": {
                        "description": "Customers retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "InternalApiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "description": "Store a newly created resource in storage.",
                "operationId": "cef1c1e98a7505ddd8d797aba6787da4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer registered successfully."
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "InternalApiKey": []
                    }
                ]
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer by ID",
                "description": "Display the specified resource.",
                "operationId": "3e5c12720ad6bcf7b3a5bb3a4a0b66e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found."
                    }
                },
                "security": [
                    {
                        "InternalApiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer",
                "description": "Remove the specified resource from storage.",
                "operationId": "ed4138d1d03b7c32c5873630c75f893e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Customer deleted successfully."
                    },
                    "404": {
                        "description": "Customer not found."
                    }
                },
                "security": [
                    {
                        "InternalApiKey": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "description": "Update the specified resource in storage.",
                "operationId": "f25886ff582e77d53e6f629201c0e760",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully."
                    },
                    "404": {
                        "description": "Customer not found."
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "InternalApiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "required": [
                    "username",
                    "last_name",
                    "first_name",
                    "email"
                ],
                "properties": {
                    "username": {
                        "description": "Customer username",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Customer last name",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "Customers first name",
                        "type": "string"
                    },
                    "email": {
                        "description": "Customer email",
                        "type": "string",
                        "format": "email"
                    },
                    "phone": {
                        "description": "Customer phone number (optional)",
                        "type": "string"
                    },
                    "is_active": {
                        "description": "Is the customer active? (optional)",
                        "type": "boolean"
                    },
                    "company_id": {
                        "description": "Company ID (optional, must exist in companies table)",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "InternalApiKey": {
                "type": "apiKey",
                "name": "X-Internal-Api-Key",
                "in": "header"
            }
        }
    },
...
}